Removes register
